refactor(domain): domain.participant에 위치되는 클래스 이동
