+++ src/main/java/domain/participant/Participants.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 public record Participants(Dealer dealer, Players players) {
