+++ src/main/java/domain/participant/PlayerName.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import exception.BlackjackException;
